requirement.txt
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fire==0.6.0
fonttools==4.51.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==5.2.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
pdf2docx==0.5.8
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
python-docx==1.1.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
termcolor==2.4.0
translate==3.6.1
typing_extensions==4.11.0
urllib3==2.2.1
